+++ lib/github_reporter.dart
@@ -1,3 +1,0 @@
-library github_reporter;
-
-export 'src/utils/gh_reporeter.dart';