+++ lib/github_snitch.dart
@@ -1,0 +1,3 @@
+library github_snitch;
+
+export 'src/gh_snitch.dart';