+++ example/lib/main.dart
@@ -126,7 +126,7 @@ class _MyHomePageState extends State<MyHomePage> {
           //Text(h),
           Text(
             '$_counter',
-            style: Theme.of(context).textTheme.headline4,
+            style: Theme.of(context).textTheme.headlineMedium,
           ),
         ],
       ),