+++ CHANGELOG.md
@@ -4,4 +4,5 @@
 * Create labels
 	- GhReporter-external for Errors not caught by Flutter Framework
 	- GhReporter-internal for Errors caught by Flutter Framework
+	- Reported by GhReporter Package for know which issues reported from this package
 * Support offline case (save locally & send later when connection exist)