+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.0.1-beta
 
 * Report bugs on github issues with specific labels,assignees, milestone
+  - Automaticlly when call `listenToExceptions` method in `main` as in example
+  - Manually with `report` method
 * Create labels
 	- GhReporter-external for Errors not caught by Flutter Framework
 	- GhReporter-internal for Errors caught by Flutter Framework